Macro tool button select works -Recent macros list now saves -Macros can now be saved/loaded from TAStudio.

+++ trunk/BizHawk.Client.Common/movie/tasproj/TasLagLog.cs
@@ -65,7 +65,7 @@ namespace BizHawk.Client.Common
 
 		public void RemoveFrom(int frame)
 		{
-			if (LagLog.Count >= frame)
+			if (LagLog.Count >= frame && frame >= 0)
 				LagLog.RemoveRange(frame, LagLog.Count - frame);
 		}
